+++ src/DependencyFactory/WardenKohanaDependencyFactory.php
@@ -9,6 +9,7 @@ namespace Ingenerator\Warden\UI\Kohana\DependencyFactory;
 
 use Ingenerator\KohanaExtras\DependencyFactory\RequestExecutorFactory;
 use Ingenerator\Warden\UI\Kohana\Controller\ChangeEmailController;
+use Ingenerator\Warden\UI\Kohana\Controller\ChangePasswordController;
 use Ingenerator\Warden\UI\Kohana\Controller\CompleteChangeEmailController;
 use Ingenerator\Warden\UI\Kohana\Controller\LoginController;
 use Ingenerator\Warden\UI\Kohana\Controller\LogoutController;
@@ -41,6 +42,16 @@ class WardenKohanaDependencyFactory
                                 '%warden.support.token_service%',
                                 '%warden.repository.user%',
                                 '%warden.user_session.session%',
+                            ],
+                        ],
+                    ],
+                    'change_password' => [
+                        '_settings' => [
+                            'class'     => \Ingenerator\Warden\Core\Interactor\ChangePasswordInteractor::class,
+                            'arguments' => [
+                                '%warden.validator.validator%',
+                                '%warden.support.password_hasher%',
+                                '%warden.repository.user%',
                             ],
                         ],
                     ],
@@ -230,6 +241,14 @@ class WardenKohanaDependencyFactory
                                 ],
                             ],
                         ],
+                        'change_password' => [
+                            '_settings' => [
+                                'class'     => \Ingenerator\Warden\UI\Kohana\View\ChangePasswordView::class,
+                                'arguments' => [
+                                    '%view.layout.default%',
+                                ],
+                            ],
+                        ],
                         'profile' => [
                             '_settings' => [
                                 'class'     => \Ingenerator\Warden\UI\Kohana\View\ProfileView::class,
@@ -256,6 +275,14 @@ class WardenKohanaDependencyFactory
                 '%warden.support.interactor_request_factory%',
                 '%warden.interactor.email_verification%',
                 '%warden.view.profile.change_email%',
+                '%warden.support.url_provider%',
+                '%warden.user_session.session%',
+                '%kohana.psr_log%',
+            ],
+            ChangePasswordController::class => [
+                '%warden.support.interactor_request_factory%',
+                '%warden.interactor.change_password%',
+                '%warden.view.profile.change_password%',
                 '%warden.support.url_provider%',
                 '%warden.user_session.session%',
                 '%kohana.psr_log%',