+++ src/Message/Profile/PasswordChangedMessage.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * @author    Andrew Coulton <andrew@ingenerator.com>
+ * @licence   proprietary
+ */
+
+namespace Ingenerator\Warden\UI\Kohana\Message\Profile;
+
+
+use Ingenerator\Pigeonhole\Message;
+use Ingenerator\Pigeonhole\Message\KohanaMessage;
+
+class PasswordChangedMessage extends KohanaMessage
+{
+    public function __construct()
+    {
+        parent::__construct(
+            'warden_flash_messages',
+            'profile.password_changed',
+            [],
+            Message::SUCCESS
+        );
+    }
+
+}