+++ src/Routing/WardenConfigBasedRouter.php
@@ -71,6 +71,11 @@ class WardenConfigBasedRouter implements UrlProvider
         return $this->requireConfigUrl('change-email');
     }
 
+    public function getChangePasswordUrl()
+    {
+        return $this->requireConfigUrl('change-password');
+    }
+
     public function getCompleteChangeEmailUrl(array $params)
     {
         return $this->requireConfigUrl('complete-change-email', $params);